Cleanup, new launcher and phoenix icon, TODOS
TODO for GamesDBContentProvider to be general for all tables in db and
for BluetoothDataExchangeActivity to visualize and select games before
and after exchange.
Bug fix for tablets: Renaming player could cause two equal players to be
in same team or in same game. A PlayerTeam (so also a PlayerDuo) do not
require different players anymore, so this fact is ignored and the game
will only be unable to load in future.